Same convergent distributed fixes as master, resolved master's way:

- ShardedModelBase: take InvalidateLayoutState(), the virtual superset of
  this branch's `CachedFullParameters = null`; keep this branch's comment
  about lazy materialization after the shard was inspected.
- ShardedOptimizerBase: comment-only conflicts, keep master's wording.
- Bucket12_DistributedTests: take master's fixture isolation
  (DistributedEnvironmentId + OwnCommunicationBackend, which owns Shutdown
  at teardown) in place of this branch's local IsolatedEnvironment(), and
  drop that helper -- both of its call sites are gone, so it would be dead
  private code.

+++ src/Preprocessing/DimensionalityReduction/SparsePCA.cs
@@ -377,7 +377,7 @@ public class SparsePCA<T> : TransformerBase<T, Matrix<T>, Matrix<T>>
                 error += diff * diff;
             }
         }
-        return error / (n * p);
+        return error / ((double)n * p);
     }
 
     private static double[,] InvertMatrix(double[,] matrix, int n)
